Implementando toString nas classes ProdutoNaoIndustrializado e ProdutoPorUnidade

+++ src/com/projeto/ProdutoQuantidadeFixa.java
@@ -44,6 +44,6 @@ public class ProdutoQuantidadeFixa extends Item {
 	}
 
 	@Override
-	public void setQuilos(int novoQuilos) {
+	public void setQuilos(double novoQuilos) {
 	}
 }
